♻️ Refractored the sorting function
Put the sorting logic in a separate file "HandleSorting" and rewrote the code to not have as myck code repeating.
