+++ src/pages/SearchPage.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from "react";
 import { JobCard } from "../components/JobCard";
 import type { Job } from "../models/Job";
-import { getJobs, sortJobsByApplicationDate, sortJobsByPublishedDateAsc, sortJobsByPublishedDateDesc } from "../services/jobService";
-import { DigiContextMenu } from "@digi/arbetsformedlingen-react";
+import { getJobs } from "../services/jobService";
+import { HandleSorting } from "../components/HandleSorting";
 
 export default function SearchPage() {
   const [jobs, setJobs] = useState<Job[]>(
@@ -20,52 +20,19 @@ export default function SearchPage() {
     getData();
   });
 
-const handleSortApplicationClick = async () => {
+  const handleSorting = async (queryParams?: string) => {
     try {
-      const data = await sortJobsByApplicationDate();
-      setJobs(data); 
-    } catch (err) {
-      console.error("Kunde inte hämta sorterade jobb", err);
+      const data = await getJobs(queryParams);
+      setJobs(data);
+    } catch {
+      throw new Error ("Kunde inte hämta sorterade jobb")
     }
-  };
-
-const handleSortPublishedClickAsc = async () => {
-    try {
-      const data = await sortJobsByPublishedDateAsc();
-      setJobs(data); 
-    } catch (err) {
-      console.error("Kunde inte hämta sorterade jobb", err);
-    }
-  };
-
-const handleSortPublishedClickDesc = async () => {
-    try {
-      const data = await sortJobsByPublishedDateDesc();
-      setJobs(data); 
-    } catch (err) {
-      console.error("Kunde inte hämta sorterade jobb", err);
-    }
-  };
-
-const menuItems = [
-    { id: 0, title: "Ansökningsdatum", onClick: handleSortApplicationClick},
-    { id: 1, title: "Publiceringsdatum (äldst först)", onClick: handleSortPublishedClickAsc},
-    { id: 2, title: "Publiceringsdatum (nyast först)", onClick: handleSortPublishedClickDesc },
-  ];
+  }
 
   return (
     <>
       <h1>Sök jobb</h1>
-
-      <DigiContextMenu
-        afTitle="Sortera efter"
-        afMenuItems={menuItems}
-        onAfChangeItem={(e: CustomEvent) => {
-          const clickedItem = e.detail.item;
-          clickedItem.onClick?.(); 
-        }}
-        >
-      </DigiContextMenu>
+      <HandleSorting onSort={handleSorting} />
 
       {jobs.map((j) => (
         <JobCard key={j.id} job={j} />
